+++ R/tweak-page.R
@@ -68,50 +68,6 @@ tweak_rmarkdown_html <- function(html, input_path, pkg = list(bs_version = 3)) {
   invisible()
 }
 
-tweak_homepage_html <- function(html,
-                                strip_header = FALSE,
-                                sidebar = TRUE,
-                                bs_version = 3,
-                                logo = NULL) {
-
-  html <- tweak_sidebar_html(html, sidebar = sidebar)
-
-  # Always remove dummy page header
-  header <- xml2::xml_find_all(html, ".//div[contains(@class, 'page-header')]")
-  if (length(header) > 0)
-    xml2::xml_remove(header, free = TRUE)
-
-  header <- xml2::xml_find_first(html, ".//h1")
-  if (strip_header) {
-    page_header <- xml2::xml_remove(header, free = TRUE)
-  } else {
-    page_header <- xml2::xml_add_parent(header, "div", class = "page-header")
-  }
-
-  if (!is.null(logo) && bs_version > 3) {
-    # Remove logo if added to h1
-    # Bare image
-    xml2::xml_remove(xml2::xml_find_all(html, ".//h1/img[contains(@src, 'logo')]"))
-
-    # Image in link
-    xml2::xml_remove(
-      xml2::xml_parent(
-        xml2::xml_find_all(html, ".//h1/a/img[contains(@src, 'logo')]")
-      )
-    )
-
-    # Add back to header
-    xml2::xml_add_sibling(page_header, "img",
-      src = logo,
-      class = "logo",
-      alt = "",
-      .where = "before"
-    )
-  }
-
-  invisible()
-}
-
 # Strip off #toc if it's not needed; easier to do this here than in js
 tweak_useless_toc <- function(html) {
   contents <- xml2::xml_find_all(html, ".//main")