+++ tests/testthat/test-tweak-page.R
@@ -1,13 +1,3 @@
-test_that("first header is wrapped in page-header div", {
-  html <- xml2::read_html('
-    <h1>First</h1>
-    <h1>Second</h1>
-  ')
-
-  tweak_homepage_html(html, bs_version = 3)
-  expect_equal(xpath_attr(html, ".//div", "class"), "page-header")
-})
-
 test_that("links to vignettes & figures tweaked", {
   html <- xml2::read_html('<body>
     <img src="vignettes/x.png" />